+++ tests/cli/test_deploy.py
@@ -2,9 +2,7 @@ import os
 from typing import ClassVar
 
 import pytest
-from fastapi.testclient import TestClient
 
-from mlem.contrib.fastapi import FastAPIServer
 from mlem.core.meta_io import MLEM_EXT
 from mlem.core.metadata import load_meta
 from mlem.core.objects import (
@@ -12,49 +10,10 @@ from mlem.core.objects import (
     DeployState,
     DeployStatus,
     MlemLink,
-    ModelMeta,
     TargetEnvMeta,
 )
 from mlem.runtime.client.base import BaseClient, HTTPClient
-from mlem.runtime.interface.base import ModelInterface
 from tests.cli.conftest import Runner
-
-
-@pytest.fixture
-def interface(model, train):
-    model = ModelMeta.from_obj(model, sample_data=train)
-    interface = ModelInterface.from_model(model)
-    return interface
-
-
-@pytest.fixture
-def client(interface):
-    app = FastAPIServer().app_init(interface)
-    return TestClient(app)
-
-
-@pytest.fixture
-def request_get_mock(mocker, client):
-    def patched_get(url, params=None, **kwargs):
-        url = url[len("http://") :]
-        return client.get(url, params=params, **kwargs)
-
-    return mocker.patch(
-        "mlem.runtime.client.base.requests.get",
-        side_effect=patched_get,
-    )
-
-
-@pytest.fixture
-def request_post_mock(mocker, client):
-    def patched_post(url, data=None, json=None, **kwargs):
-        url = url[len("http://") :]
-        return client.post(url, data=data, json=json, **kwargs)
-
-    return mocker.patch(
-        "mlem.runtime.client.base.requests.post",
-        side_effect=patched_post,
-    )
 
 
 @pytest.mark.usefixtures("request_get_mock", "request_post_mock")
