+++ tests/runtime/test_client.py
@@ -2,17 +2,13 @@ import platform
 
 import numpy as np
 import pytest
-from fastapi.testclient import TestClient
 
 from mlem.constants import PREDICT_ARG_NAME, PREDICT_METHOD_NAME
-from mlem.contrib.fastapi import FastAPIServer
 from mlem.contrib.numpy import NumpyNdarrayType
 from mlem.core.dataset_type import DatasetAnalyzer
 from mlem.core.errors import WrongMethodError
 from mlem.core.model import Argument, Signature
-from mlem.core.objects import ModelMeta
 from mlem.runtime.client.base import HTTPClient
-from mlem.runtime.interface.base import ModelInterface
 
 
 @pytest.fixture
@@ -28,43 +24,6 @@ def signature(train):
         args=[Argument(name=PREDICT_ARG_NAME, type_=data_type)],
         returns=returns_type,
         **kwargs,
-    )
-
-
-@pytest.fixture
-def interface(model, train):
-    model = ModelMeta.from_obj(model, sample_data=train)
-    interface = ModelInterface.from_model(model)
-    return interface
-
-
-@pytest.fixture
-def client(interface):
-    app = FastAPIServer().app_init(interface)
-    return TestClient(app)
-
-
-@pytest.fixture
-def request_get_mock(mocker, client):
-    def patched_get(url, params=None, **kwargs):
-        url = url[len("http://") :]
-        return client.get(url, params=params, **kwargs)
-
-    return mocker.patch(
-        "mlem.runtime.client.base.requests.get",
-        side_effect=patched_get,
-    )
-
-
-@pytest.fixture
-def request_post_mock(mocker, client):
-    def patched_post(url, data=None, json=None, **kwargs):
-        url = url[len("http://") :]
-        return client.post(url, data=data, json=json, **kwargs)
-
-    return mocker.patch(
-        "mlem.runtime.client.base.requests.post",
-        side_effect=patched_post,
     )
